Bf16 dummy

+++ torch/csrc/jit/passes/common_subexpression_elimination.cpp
@@ -1,19 +1,16 @@
-#include <torch/csrc/jit/ir.h>
-
-#include <algorithm>
-#include <unordered_map>
-
-#include <torch/csrc/jit/assertions.h>
-#include <torch/csrc/jit/interned_strings.h>
-#include <torch/csrc/jit/passes/alias_analysis.h>
 #include <torch/csrc/jit/passes/common_subexpression_elimination.h>
+
+#include <torch/csrc/jit/ir.h>
+#include <torch/csrc/jit/jit_log.h>
 #include <torch/csrc/jit/node_hashing.h>
-#include <torch/csrc/utils/functional.h>
-#include <torch/csrc/utils/hash.h>
+#include <torch/csrc/jit/passes/alias_analysis.h>
+
+#include <unordered_map>
 
 namespace torch {
 namespace jit {
 namespace {
+
 // The function implements common subexpression elimination.
 // Since the nodes are visited in topological order, one pass is enough.
 void EliminateCommonSubexpression(
@@ -21,11 +18,18 @@ void EliminateCommonSubexpression(
     const AliasDb& aliasDb,
     std::function<Node*(Node*)> parent_lookup_fn) {
   std::unordered_set<Node*, HashNode, EqualNode> subexprs;
-  for (auto it = block->nodes().begin(); it != block->nodes().end(); ++ it) {
+  for (auto it = block->nodes().begin(); it != block->nodes().end(); ++it) {
     auto node = *it;
-    if (node->kind() == prim::PythonOp || node->kind() == prim::Print ||
-        node->kind() == aten::warn || node->isNondeterministic() ||
-        aliasDb.hasWriters(node) || aliasDb.hasWildcard(node)) {
+    if (node->hasSideEffects()) {
+      GRAPH_DEBUG("Node was skipped due to side effects:\n", *node);
+      continue;
+    }
+    if (node->isNondeterministic()) {
+      GRAPH_DEBUG("Node was skipped due to its non determinism:\n", *node);
+      continue;
+    }
+    if (aliasDb.hasWriters(node)) {
+      GRAPH_DEBUG("Node was skipped due to alias analysis result:\n", *node);
       // Do NOT have enough information to do CSE on these nodes.
       continue;
     }
@@ -48,7 +52,16 @@ void EliminateCommonSubexpression(
 
     // Check for CSE opportunities in the parent block.
     auto parent_lookup = parent_lookup_fn(node);
+    auto g_out = node->owningGraph()->outputs();
     if (parent_lookup) {
+      // since the graph outputs may be mutated after they are returned,
+      // don't introduce new aliasing among graph outputs
+      if (aliasDb.mayContainAlias(node->outputs(), g_out) &&
+          aliasDb.mayContainAlias(parent_lookup->outputs(), g_out)) {
+        continue;
+      }
+
+      GRAPH_UPDATE("Replacing\n", *node, "with\n", *parent_lookup);
       node->replaceAllUsesWith(parent_lookup);
       it.destroyCurrent();
       continue;
@@ -59,16 +72,26 @@ void EliminateCommonSubexpression(
     if (!subit.second) {
       // Subexpression exists, replace the uses of node, and destroy it.
       auto existing = *subit.first;
+
+      // don't introduce new aliasing among graph outputs
+      if (aliasDb.mayContainAlias(
+              node->outputs(), node->owningGraph()->outputs()) &&
+          aliasDb.mayContainAlias(existing->outputs(), g_out)) {
+        continue;
+      }
+
+      GRAPH_UPDATE("Replacing\n", *node, "with\n", *existing);
       node->replaceAllUsesWith(existing);
       // Destroy the node.
       it.destroyCurrent();
     }
   }
 }
-}
+} // namespace
 
 void EliminateCommonSubexpression(std::shared_ptr<Graph>& graph) {
-  const auto aliasDb = AliasAnalysis(graph);
+  AliasDb aliasDb(graph);
+  GRAPH_DUMP("Before CSE", graph);
   EliminateCommonSubexpression(
       graph->block(), aliasDb, [](Node*) { return nullptr; });
 }
